git.py: properly handle single-character flags with arguments
"git diff -U5" is correct while "git diff -U 5" is not.
ditto for "git log -U5", "git commit -F/some/path", etc.

The original version of transform_kwargs was setting up the
command line arguments such that single-character flags with
arguments were not understood by git.

This changes transform_kwargs so that the flag and argument
are both part of the same string for single-character flags.

Signed-off-by: David Aguilar <davvid@gmail.com>