cmd: properly handle cwd for repos with work trees
This is a fix on top of Govind's latest performance improvement.

self._cwd was always set to self._git_dir which means
a lot of commands that require work trees were not available
to GitPython.  Execute now uses self._cwd which is equal to
self._git_dir by default, and self.get_work_tree() if a work tree
exists.

Signed-off-by: David Aguilar <davvid@gmail.com>