cmd: better support for bare repositories
In order to avoid the expense of parsing .git/config
just to know whether or not a repository is bare
at __init__ time, we just pass an optional flag to
Git.__init__():
	bare_repo with a default value of False.

Repo.init_bare() was updated to pass this flag.

We could have an optional Git.read_bare_status() function
that does the expensive lookup.  Then, users can optionally
call it at runtime if they really need to know whether or not
a repository is bare.  That seems like a decent tradeoff between
speed, correctness, and common use cases.

Signed-off-by: David Aguilar <davvid@gmail.com>