Fixed packaging issue caused by ez_setup.py
ez_setup.py is not packaged with the rest of the code, this causes
`easy_install gitpython` to fail with an ImportError.

ez_setup is useful when setuptools are not present in the pythonpath.
(cherry picked from commit e7c37c6e0b12d2b4f5001fb3ecfc250a682f6d57)