Make message the full message and add a summary property
Adds a summary property to the Commit object which returns just the first line of the commit message and makes the message property contain the entire commit message (previously the message property only contained the first line of the commit message).  This breaks backwards compatibility a little in that the message property now contains a different value but previously there was no way to access the entire commit message from the Commit object and this is in keeping with git vocabulary, where message generally refers to the entire commit message.
(cherry picked from commit 9d2962d8306c894d4cca55bab551677b92d96352)