Verified that it apparently is impossible to add empty submodules using git-python.
This is the case with `git submodule add` as well. This makes sense as
an empty git repository doesn't have a commit, which needs to be specified
as SHA in the parent repositories tree entry for the respective submodule.

When manually adding the empty submodule to the .gitmodules file, git-python
will throw another error related to the inability to find the submodule
in the index.

Even if an iteration would be possible, git-python would now throw
a BadName exception, which clearly indicates that the 'HEAD' revision
is invalid (as it doesn't point to any commit).

Fixes #152
Fixes #105