fix(remote): don't close stdout on fetch/pull
Reverted changes of `fe2fbc5~2`.

This caused `git-pull` to error, which now actually results in a fatal
error while fetching or pulling. Previously we simply didn't check
for this issue.

Now we are back to a `poll` based or threaded concurrent reading from
stdout and stderr to prevent a git process deadlock, and the
aforementioned error.

Related to #297