fix(travis): assure private key has access
Previously travis would fail to clone the special testing repository
as SSH access requires a deploy key to be set. The latter has been
configured and is now added in clear-text.

This is save as it is only good for read-only access to a test
repository, mis-use cannot happen that way.