fix(repo): fail loudly if worktrees are used
As GitPython is in maintenance mode, there will be no new features.
However, I believe it's good idea to explicitly state we do not support
certain things if this is the case.

Therefore, when worktrees are encountered, we will throw an specific
exception to indicate that.

The current implementation is hacky to speed up development,
and increases the risk of failing due to false-positive worktree
directories.

Related to #344