Support repeated kwargs
Some Git command line options are allowed to be repeated multiple times.
Examples of this are the -C flag which may occur more than once to
"strengthen" its effect, or the -L flag on Git blames, to select
multiple blocks of lines to blame.

    $ git diff -C -C HEAD~1 HEAD
    $ git blame -L 1-3 -L 12-18 HEAD -- somefile.py

This patch supports passing a list/tuple as the value part for kwargs,
so that the generated Git command contain the repeated options.