Travis should now be able to test tags
It's just a guess, maybe we are lucky.
The original problem is that travis checks out tags without branches,
and thus checking out master does only work if travis runs on master.

With tags, it will only heckout and locally know the tag in question.
The changes should allow it to retry and create the master branch
instead.