doc(README): add basic contribution guidelines
The main point is that from now on, tests are required to add
new features. If the fix is minor enough, not having a test
is probably alright.

That distinction is not represented in the contribution guide
as more tests are better - people should prefer to have a test
whenever they contribute anything.

My motivation to finally do this is the sad realization that
I grow too unconfident about the quality of some contributions
without having tests that proof they are valid. It's not enough
to not break anything that exists, as the current test-suite is
certainly not perfect either.