Fix exception on import in MacOS
This is related to my fix in #658. Apparently, MacOS adds a git
executable that is just a stub which displays an error. This gets past
the try/except I added in #658, and allows all of the GitPython
components to be imported, but since the executable is not *actually*
git, it results in an exception when ``refresh()`` attemepts to run a
``git version``.