Dedent code blocks in tutorial.
I found the extra 8 spaces at the start of the examples in the tutorial
to be distracting. The Sphinx dedent option removes these extra spaces
from the rendered code blocks.

I also got a warning about the shell code example not being lexed as
Python, so I converted this to an explicit shell code block.