Get system user id in a lazy manner
Calling getpass.getuser may lead to breakage in environments where there
is no entries in the /etc/passwd file for the current user.

Setting the environment variables for the git user configurations should
prevents GitPython from using values from /etc/passwd. However, doing so
will not prevent reading /etc/passwd and looking for an entry with the
current user UID.

This patch changes the behavior described above so GitPython will
perform a lazy evaluation of /etc/passwd, only doing so when the
environment variables for the git user configuration are not available.

Signed-off-by: Athos Ribeiro <athos@redhat.com>