First attempt to fix failing test of #1103
However, the test asserts on the provided context to be correct,
which is hard to do in this branch while it's easy to doubt the value
of this.

Lastly, there seems to be no way to ignore errors in `git` without
muting all output, which is in fact parsed.

Maybe it's possible to ignore errors while parsing the new kind of
error message.