fix(fetch): use the correct FETCH_HEAD from within a worktree
FETCH_HEAD is one of the symbolic references local to the current
worktree and as such should _not_ be looked up in the 'common_dir'. But
instead of just hard coding the "right thing" (git_dir) lets defer this
to the SymbolicReference class which already contains this knowledge in
its 'abspath' property.