improve index mode for files with executable bit
The fix for #430 in bebc4f56 (Use correct mode for executable files,
2016-05-19) is incomplete.  It fails (in most cases) when files have
modes which are not exactly 0644 or 0755.

Git only cares whether the executable bit is set (or not).  Ensure the
mode we set for the index is either 100644 or 100755 based on whether
the executable bit is set for the file owner.  Do this similarly to how
upstream git does it in cache.h¹.

Add a test covering various file modes to help catch regressions.

Fixes #1253

¹ https://github.com/git/git/blob/v2.31.1/cache.h#L247