Add test asserting that get_values works by itself
As described in #1534, this test will fail in main with a KeyError
about a missing section name even though the named sections do exist
within the config file.

This test will pass in the branch associated with #1535. This test
should keep the improved behavior healthy as the code evolves by
preventing future developers from removing the eager loading of
sections within get_values().