Only set safe.directory on Cygwin (which needs it)
This stops setting the current directory as an explicit safe
directory on CI for non-Windows systems, where this is not needed
because the repository has the ownership Git expects. The step name
is updated accordingly to reflect its now narrower purpose.

This also adds shell quoting to $(pwd) in the Cygwin workflow. In
practice, on CI, the path is very unlikely to contain whitespace,
but double-quoting $ expansions on which splitting and globbing are
unwanted is more robust and better expresses intent. This also has
the benefit that users who use the CI workflows as a guide to
commands they run locally, where on Windows they may very well have
spaces somewhere in this absolute path, will use a correct command.