Set protocol.file.allow only in tests that need it
Instead of setting environment variables just on CI and for the
the entire pytest command, this has the two test cases that need
protocol.file.allow to be set to "always" (instead of "user") set
them, via a shared fixture, just while those tests are running.

Both on CI and for local test runs, this makes it no longer
necessary to set this in a global configuration or through
environment variables, reducing the setup needed to run the tests.