Redesign new decorator to better separate concerns
_allow_file_protocol was effectively a _patch_git_config fixture,
being no no shorter, simpler, or clearer by hard-coding the
specific name and value to patch. So this changes it to be that.

As a secondary issue, it previously was called with no arguments,
then that would be used as a decorator. That was unintutive and it
was easy to omit the parentheses accidentally. This resolves that.