Use "python" in the virtual env, "python3" outside
This changes the build command to run with "python" when in a
virtual environment, since all virtual environments support this
even when "python" outside it is absent or refers to the wrong
version.

On Windows, virtual environments don't contain a python3 command,
but a global python3 command may be present, so the errors are
confusing. This fixes that by avoiding such errors altogether.