Fix non-venv branch always failing
cc202cc put an end to the problem where, when run outside a virtual
environment, it would always "succeed", because all "|| echo ..."
required to succeed was for the echo command reporting the error to
succeed. Unfortunately, that commit created the oppposite problem,
causing that case to always fail! This commit fixes it, so it fails
when there is an error, and succeeds when there is no error.